Optionally, use the analog comparator as the frequency counter.
The problem is that each ADC sample takes something like 70 µs even if we crank the ADC clock up to 256 kHz. That means that we can skip entirely over state changes when the duty cycle is near one end of the range or the other. Since the AC is interrupt driven, this can't happen.

But this requires hardware support - we need to give the AC a reference of something close to the 0 volt reading (2.5v is close enough) because the internal reference isn't a useful level.