1. Up the sample rate. 2. zero duty cycle when zero frequency
We used to want a long sample period so we could make sure we exhaustively captured data points across the entire period. Now that we're using the AC and timer to measure duty cycle, we don't care quite so much, so we can report more frequently.

If there are no hi-lo transitions at all, then the hi and lo periods are just going to be what they last were, which makes no sense.